[wip] Reorganize files.
